d/rules configure: drop SYSTEM_TYPE=
This is used as part of the labels printed by various utils like
my_print_defaults. If not set it will be generated by cmake. Since
cmake will generate this, no need to have it set explicitly. This will
make the label more generic, but we have no need to have it specific.