d/mysql-server-5.7.postinst: Use new function to set root pw
Using the more general function for resetting the password lets us
remove the old one.

The old password setting code checked for failure and printed a debconf
message. Now that we're using --init-file and don't need login access,
this is not likely to fail so the debconf note is no longer necessary.