mysql-server-5.7: depend on apparmor (>= 2.10)
This is required in Ubuntu to fix the upgrade path from Trusty.
Otherwise the new mysqld fails to start because the new AppArmor profile
contains Unix socket restrictions which the older apparmor userspace
tools cannot parse.

This is not believed to affect Debian because jessie has AppArmor 2.9
which has the required support. A versioned dependency on 2.9 may do,
but I have not tested this and I know 2.10 will work.